Implementando un poco inventario( me falta un monton no pero bueno)

+++ src/main/java/com/project/minimercado/model/bussines/Producto.java
@@ -5,7 +5,7 @@ import lombok.Getter;
 import lombok.Setter;
 import org.hibernate.annotations.ColumnDefault;
 
-import java.math.BigDecimal;
+
 import java.util.LinkedHashSet;
 import java.util.Set;
 
@@ -30,10 +30,10 @@ public class Producto {
     private String descripcion;
 
     @Column(name = "precio_compra", nullable = false, precision = 10, scale = 2)
-    private BigDecimal precioCompra;
+    private Double precioCompra;
 
     @Column(name = "precio_venta", nullable = false, precision = 10, scale = 2)
-    private BigDecimal precioVenta;
+    private Double precioVenta;
 
     @ManyToOne(fetch = FetchType.LAZY, optional = false)
     @JoinColumn(name = "id_categoria", nullable = false)
